Add displayExchangeTable to table module
We now call it from its module in exchange.js by solving everything related to it directly in its module.
